Preliminary fix for #15.
Time complexity used to be O(N)+O(M) where N is the number of data
samples per chart width and M is the chart width in pixels (with N being
much more than M), now it is just O(M).

It remains to be seen whether the visuals are satisfactory with new
averaging algorithm, and some redundant code removal is required.